+++ CHANGELOG.md
@@ -8,10 +8,11 @@ The releases of major and minor versions contain an overview of changes since th
 Version 1.6.x
 -------------
 
-## Version 1.6.1 (??)
+## Version 1.6.2 (2020/09)
 - Prism program simplification improved.
 - Revamped implementation of long-run-average algorithms, including scheduler export for LRA properties on Markov automata.
-- Support for step-bounded properties of the form ... [F[x,y] ... ] for DTMCs and MDPs (sparse engine). 
+- Support for step-bounded properties of the form ... [F[x,y] ... ] for DTMCs and MDPs (sparse engine).
+- Renamed portfolio engine to automatic 
 - `storm-dft`: Fix for relevant events when using symmetry reduction.
 - `storm-pomdp`: Fix for --transformsimple and --transformbinary when used with until formulae.
 - `storm-pomdp`: POMDPs can be parametric as well.